[REFACTOR] 리팩토링 준비와 테스트 클래스 생성
