+++ src/test/java/racing/domain/car/entity/BasicCarTest.java
@@ -1,12 +1,14 @@
-package racing.domain.car;
+package racing.domain.car.entity;
 
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
-import racing.domain.Location;
-import racing.domain.Name;
-import racing.domain.fuel.BasicFuel;
+import racing.domain.car.entity.BasicCar;
+import racing.domain.car.entity.Car;
+import racing.domain.car.vo.Location;
+import racing.domain.car.vo.Name;
+import racing.domain.car.vo.fuel.BasicFuel;
 
 import static org.assertj.core.api.Assertions.*;
