+++ src/test/java/racing/domain/car/vo/LocationTest.java
@@ -1,8 +1,9 @@
-package racing.domain;
+package racing.domain.car.vo;
 
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.ValueSource;
+import racing.domain.car.vo.Location;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
