+++ src/test/java/racing/domain/game/entity/RacingGameTest.java
@@ -1,0 +1,7 @@
+package racing.domain.game.entity;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+class RacingGameTest {
+
+}
