+++ src/test/java/racing/domain/game/vo/turn/TurnTest.java
@@ -1,0 +1,7 @@
+package racing.domain.game.vo.turn;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+class TurnTest {
+
+}
