+++ src/test/java/racing/domain/game/vo/turn/TurnsTest.java
@@ -1,0 +1,7 @@
+package racing.domain.game.vo.turn;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+class TurnsTest {
+
+}
